[Example] Do not allow productionBroker to getAmount of non-existent order
ProductionBroker example works with one production station now. 
I regenerated the code and added the makefile for the broker to c-src, since it needs
unqlite. 
ProductionBroker still tried to getAmount for orderId -6 if no order was found. It now 
checks and assigns -1 to amount and ingredient if there is no order.